+++ example/product.dart
@@ -1,0 +1,3 @@
+class Product {
+  late String id;
+}