+++ lib/src/models/paginated_items_response.dart
@@ -1,40 +1,52 @@
+/// The response object that carries the list of items and handles pagination
+/// internally. The [paginationKey] is optional and can be of any type. If not passed,
+/// it is assumed that the API does not support pagination.
+///
+///
+/// The [idGetter] should be passed when receiving the response from the API, it
+/// is required for functions like `updateItem`, `findByUid`
+/// and avoiding duplication of items in list (compares id).
 class PaginatedItemsResponse<T> {
+  /// List of items of type [T]
   List<T>? items;
+
+  /// The pagination key. Can be null.
   dynamic paginationKey;
-  String Function(T)? uidGetterFn;
 
-  bool get hasMoreData {
-    if (paginationKey == null) {
-      return false;
-    } else {
-      return paginationKey!.length != 2;
-    }
-  }
+  /// ID getter for the object of type [T].
+  String Function(T)? _idGetterFn;
 
-  T? findByUid(String uid) {
-    final idx = items?.indexWhere((e) => uid == uidGetterFn!(e));
+  /// If pagination supported, check if there is more data that can be loaded.
+  bool get hasMoreData => paginationKey != null;
+
+  /// Find an object by [id].
+  T? findByUid(String id) {
+    final idx = items?.indexWhere((e) => id == _idGetterFn!(e));
     if (idx != null && idx != -1) return items?[idx];
   }
 
+  /// constructor
   PaginatedItemsResponse({
     String Function(T)? idGetter,
     Iterable<T>? listItems,
-    dynamic lastKey,
+    dynamic paginationKey,
   }) {
-    uidGetterFn ??= idGetter;
-    _update(listItems, lastKey);
+    _idGetterFn ??= idGetter;
+    _update(listItems, paginationKey);
   }
 
   void update(PaginatedItemsResponse<T>? res) {
     if (res != null) _update(res.items, res.paginationKey);
   }
 
+  /// update a specific item with uid, or add if does not exists according to
+  /// [addIfDoesNotExist].
   void updateItem(
     String itemUid,
     T? item, {
     bool addIfDoesNotExist = false,
   }) {
-    final idx = items!.indexWhere((e) => itemUid == uidGetterFn!(e));
+    final idx = items!.indexWhere((e) => itemUid == _idGetterFn!(e));
     if (idx != -1) {
       if (item == null) {
         items!.removeAt(idx);
@@ -46,16 +58,18 @@ class PaginatedItemsResponse<T> {
     }
   }
 
-  void _update(Iterable<T>? listItems, dynamic lastKey) {
+  /// Append items to the list, after a successful fetch from the API.
+  void _update(Iterable<T>? listItems, dynamic key) {
     items ??= [];
     if (listItems != null) {
       for (final item in listItems) {
-        updateItem(uidGetterFn!(item), item, addIfDoesNotExist: true);
+        updateItem(_idGetterFn!(item), item, addIfDoesNotExist: true);
       }
     }
-    paginationKey = lastKey?.toString();
+    paginationKey = key;
   }
 
+  /// Clear the contents.
   void clear() {
     items = null;
     paginationKey = null;