+++ lib/src/pagination_items_state_handler.dart
@@ -1,10 +1,18 @@
 import 'package:flutter/material.dart';
-import 'package:paginated_items_builder/src/models/paginated_items_response.dart';
+import 'package:paginated_items_builder/paginated_items_builder.dart';
 
+/// Do not have a controller for a list of items? Or do not want to create one just because of
+/// one time use, use [PaginationItemsStateHandler], it handles the state internally and
+/// wraps above a [PaginatedItemsBuilder].
+///
+/// The [itemsBuilder] provides the required arguments needed by the [PaginatedItemsBuilder].
 class PaginationItemsStateHandler<T> extends StatefulWidget {
-  static const id = 'PaginationItemsStateHandler';
+  /// Pass in a function that calls the API and returns a [PaginatedItemsResponse].
+  final Future<PaginatedItemsResponse<T>?> Function(String? paginationKey)
+      pageFetchData;
 
-  final Future<PaginatedItemsResponse<T>?> Function(String? paginationKey) pageFetchData;
+  /// Callback method that usually should return a [PaginatedItemsBuilder] and
+  /// pass the [response] and [fetchPageData] params to the builder.
   final Widget Function(
     PaginatedItemsResponse<T>? response,
     Future<void> Function(bool) fetchPageData,