+++ CHANGELOG.md
@@ -1,3 +1,5 @@
-## 0.0.1
+## 0.0.1 - 10/01/2022
 
-* Easier to display items in a list/grid view from your controllers directly or handling state internally with support for pagination.
+* Easier to display items in a list/grid view from your controllers directly or handling state
+  internally with support for pagination. Saves the results in state to avoid unnecessary api calls
+  everytime screen is pushed.